requirements.txt
Pillow>=10.0.0
beautifulsoup4
lxml